requirements.txt
openai~=1.13.3
python-dotenv~=1.0.1
langchain~=0.1.10
pinecone-client~=3.1.0
langchain_openai
langchain_pinecone